fix config validation test and extract it in a separate file
